requirements.txt
PyMuPDF
Wand
pillow
MagickWand
